feat: normalize nested/recursive usage
