+++ internal/graph/check_fast_path_test.go
@@ -3,14 +3,19 @@ package graph
 import (
 	"context"
 	"errors"
+	"fmt"
+	"strconv"
+	"sync"
 	"testing"
 
+	"github.com/emirpasic/gods/sets/hashset"
 	"github.com/oklog/ulid/v2"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/goleak"
 	"go.uber.org/mock/gomock"
 
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
+	parser "github.com/openfga/language/pkg/go/transformer"
 
 	"github.com/openfga/openfga/internal/concurrency"
 	"github.com/openfga/openfga/internal/mocks"
@@ -924,5 +929,564 @@ func TestFastPathDifference(t *testing.T) {
 				require.Equal(t, tt.expected, ids)
 			})
 		}
+	})
+}
+
+func TestBreadthFirstRecursiveMatch(t *testing.T) {
+	t.Cleanup(func() {
+		goleak.VerifyNone(t)
+	})
+
+	tests := []struct {
+		name                 string
+		currentLevelUsersets *hashset.Set
+		usersetFromUser      *hashset.Set
+		readMocks            [][]*openfgav1.Tuple
+		expectedOutcomes     []checkOutcome
+	}{
+		{
+			name:                 "empty_userset",
+			currentLevelUsersets: hashset.New(),
+			usersetFromUser:      hashset.New(),
+			expectedOutcomes:     []checkOutcome{},
+		},
+		{
+			name:                 "duplicates_no_match_no_recursion",
+			currentLevelUsersets: hashset.New("group:1", "group:2", "group:3", "group:1"),
+			usersetFromUser:      hashset.New(),
+			readMocks: [][]*openfgav1.Tuple{
+				{{}},
+				{{}},
+				{{}},
+			},
+			expectedOutcomes: []checkOutcome{
+				{resp: &ResolveCheckResponse{
+					Allowed: false,
+				}},
+			},
+		},
+		{
+			name:                 "duplicates_no_match_with_recursion",
+			currentLevelUsersets: hashset.New("group:1", "group:2", "group:3"),
+			usersetFromUser:      hashset.New(),
+			readMocks: [][]*openfgav1.Tuple{
+				{{Key: tuple.NewTupleKey("group:1", "parent", "group:3")}},
+				{{Key: tuple.NewTupleKey("group:3", "parent", "group:2")}},
+				{{Key: tuple.NewTupleKey("group:2", "parent", "group:1")}},
+			},
+			expectedOutcomes: []checkOutcome{
+				{resp: &ResolveCheckResponse{
+					Allowed: false,
+				}},
+				{resp: &ResolveCheckResponse{
+					Allowed: false,
+				}},
+			},
+		},
+		{
+			name:                 "duplicates_match_with_recursion",
+			currentLevelUsersets: hashset.New("group:1", "group:2", "group:3"),
+			usersetFromUser:      hashset.New("group:4"),
+			readMocks: [][]*openfgav1.Tuple{
+				{{Key: tuple.NewTupleKey("group:1", "parent", "group:3")}},
+				{{Key: tuple.NewTupleKey("group:2", "parent", "group:1")}},
+				{{Key: tuple.NewTupleKey("group:3", "parent", "group:4")}},
+			},
+			expectedOutcomes: []checkOutcome{
+				{resp: &ResolveCheckResponse{
+					Allowed: true,
+				}},
+			},
+		},
+		{
+			name:                 "duplicates_match_with_recursion",
+			currentLevelUsersets: hashset.New("group:1", "group:2", "group:3"),
+			usersetFromUser:      hashset.New("group:4"),
+			readMocks: [][]*openfgav1.Tuple{
+				{{Key: tuple.NewTupleKey("group:1", "parent", "group:3")}},
+				{{Key: tuple.NewTupleKey("group:2", "parent", "group:1")}},
+				{{Key: tuple.NewTupleKey("group:3", "parent", "group:4")}},
+			},
+			expectedOutcomes: []checkOutcome{
+				{resp: &ResolveCheckResponse{
+					Allowed: true,
+				}},
+			},
+		},
+		{
+			name:                 "no_duplicates_no_match_counts",
+			currentLevelUsersets: hashset.New("group:1", "group:2", "group:3"),
+			usersetFromUser:      hashset.New(),
+			readMocks: [][]*openfgav1.Tuple{
+				{{Key: tuple.NewTupleKey("group:1", "parent", "group:4")}},
+				{{Key: tuple.NewTupleKey("group:2", "parent", "group:5")}},
+				{{Key: tuple.NewTupleKey("group:3", "parent", "group:6")}},
+				{{Key: tuple.NewTupleKey("group:6", "parent", "group:9")}},
+				{{Key: tuple.NewTupleKey("group:7", "parent", "group:10")}},
+				{{Key: tuple.NewTupleKey("group:8", "parent", "group:11")}},
+				{{}},
+				{{}},
+				{{}},
+			},
+			expectedOutcomes: []checkOutcome{
+				{resp: &ResolveCheckResponse{
+					Allowed: false,
+				}},
+				{resp: &ResolveCheckResponse{
+					Allowed: false,
+				}},
+				{resp: &ResolveCheckResponse{
+					Allowed: false,
+				}},
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			ctrl := gomock.NewController(t)
+			defer ctrl.Finish()
+			ds := mocks.NewMockRelationshipTupleReader(ctrl)
+
+			storeID := ulid.Make().String()
+
+			for _, mock := range tt.readMocks {
+				ds.EXPECT().Read(gomock.Any(), storeID, gomock.Any(), gomock.Any()).Times(1).Return(storage.NewStaticTupleIterator(mock), nil)
+			}
+
+			model := parser.MustTransformDSLToProto(`
+				model
+					schema 1.1
+				type user
+				type group
+					relations
+						define member: [user] or member from parent
+						define parent: [group]
+				`)
+
+			req := &ResolveCheckRequest{
+				StoreID:              storeID,
+				AuthorizationModelID: ulid.Make().String(),
+				TupleKey:             tuple.NewTupleKey("group:3", "member", "user:maria"),
+				RequestMetadata:      NewCheckRequestMetadata(),
+			}
+
+			ts, err := typesystem.New(model)
+			require.NoError(t, err)
+			ctx := context.Background()
+			ctx = storage.ContextWithRelationshipTupleReader(ctx, ds)
+			ctx = typesystem.ContextWithTypesystem(ctx, ts)
+
+			checker := NewLocalChecker()
+			mapping := &recursiveMapping{
+				kind:             TTUKind,
+				tuplesetRelation: "parent",
+			}
+			checkOutcomeChan := make(chan checkOutcome, 100) // large buffer since there is no need to concurrently evaluate partial results
+			checker.breadthFirstRecursiveMatch(ctx, req, mapping, &sync.Map{}, tt.currentLevelUsersets, tt.usersetFromUser, checkOutcomeChan)
+
+			collectedOutcomes := make([]checkOutcome, 0)
+			for outcome := range checkOutcomeChan {
+				collectedOutcomes = append(collectedOutcomes, outcome)
+			}
+			require.Equal(t, tt.expectedOutcomes, collectedOutcomes)
+		})
+	}
+}
+
+func TestRecursiveTTUFastPath(t *testing.T) {
+	t.Cleanup(func() {
+		goleak.VerifyNone(t)
+	})
+
+	model := parser.MustTransformDSLToProto(`
+				model
+					schema 1.1
+				type user
+				type group
+					relations
+						define member: [user] or member from parent
+						define parent: [group]
+				`)
+	tests := []struct {
+		name                            string
+		readStartingWithUserTuples      []*openfgav1.Tuple
+		readStartingWithUserTuplesError error
+		readTuples                      [][]*openfgav1.Tuple
+		readTuplesError                 error
+		expected                        *ResolveCheckResponse
+		expectedError                   error
+	}{
+		{
+			name: "happy_case",
+			readStartingWithUserTuples: []*openfgav1.Tuple{
+				{
+					Key: tuple.NewTupleKey("group:1", "member", "user:maria"),
+				},
+			},
+			readTuples: [][]*openfgav1.Tuple{
+				{
+					{
+						Key: tuple.NewTupleKey("group:2a", "parent", "group:1a"),
+					},
+					{
+						Key: tuple.NewTupleKey("group:3", "parent", "group:2a"),
+					},
+					{
+						Key: tuple.NewTupleKey("group:3", "parent", "group:2"),
+					},
+				},
+			},
+			expected: &ResolveCheckResponse{
+				Allowed: true,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+
+			ctrl := gomock.NewController(t)
+			defer ctrl.Finish()
+			ds := mocks.NewMockRelationshipTupleReader(ctrl)
+			storeID := ulid.Make().String()
+			ds.EXPECT().ReadStartingWithUser(gomock.Any(), storeID, storage.ReadStartingWithUserFilter{
+				ObjectType: "group",
+				Relation:   "member",
+				UserFilter: []*openfgav1.ObjectRelation{{Object: "user:maria"}},
+				ObjectIDs:  nil,
+			}, gomock.Any()).MaxTimes(1).Return(storage.NewStaticTupleIterator(tt.readStartingWithUserTuples), tt.readStartingWithUserTuplesError)
+
+			for _, tuples := range tt.readTuples {
+				ds.EXPECT().Read(gomock.Any(), storeID, gomock.Any(), gomock.Any()).MaxTimes(1).Return(storage.NewStaticTupleIterator(tuples), tt.readTuplesError)
+			}
+			ts, err := typesystem.New(model)
+			require.NoError(t, err)
+			rel, err := ts.GetRelation("group", "parent")
+			require.NoError(t, err)
+
+			req := &ResolveCheckRequest{
+				StoreID:              storeID,
+				AuthorizationModelID: ulid.Make().String(),
+				TupleKey:             tuple.NewTupleKey("group:3", "member", "user:maria"),
+				RequestMetadata:      NewCheckRequestMetadata(),
+			}
+			ctx := context.Background()
+			ctx = storage.ContextWithRelationshipTupleReader(ctx, ds)
+			ctx = typesystem.ContextWithTypesystem(ctx, ts)
+			checker := NewLocalChecker()
+			result, err := checker.recursiveTTUFastPath(ctx, req, rel.GetRewrite(), storage.NewStaticTupleKeyIterator([]*openfgav1.TupleKey{{Object: "group:2", Relation: "parent", User: "group:1"}}))
+			require.Equal(t, tt.expectedError, err)
+			require.Equal(t, tt.expected.GetAllowed(), result.GetAllowed())
+			require.Equal(t, tt.expected.GetResolutionMetadata(), result.GetResolutionMetadata())
+		})
+	}
+}
+
+func TestRecursiveUsersetFastPath(t *testing.T) {
+	t.Cleanup(func() {
+		goleak.VerifyNone(t)
+	})
+	tests := []struct {
+		name                            string
+		readStartingWithUserTuples      []*openfgav1.Tuple
+		readStartingWithUserTuplesError error
+		readUsersetTuples               [][]*openfgav1.Tuple
+		readUsersetTuplesError          error
+		expected                        *ResolveCheckResponse
+		expectedError                   error
+	}{
+		{
+			name:                       "no_user_assigned_to_group",
+			readStartingWithUserTuples: []*openfgav1.Tuple{},
+			readUsersetTuples: [][]*openfgav1.Tuple{
+				{
+					{
+						Key: tuple.NewTupleKey("group:1", "member", "group:3#member"),
+					},
+				},
+				{},
+			},
+			expected: &ResolveCheckResponse{
+				Allowed: false,
+			},
+		},
+		{
+			name: "user_assigned_to_first_level_sub_group",
+			readStartingWithUserTuples: []*openfgav1.Tuple{
+				{
+					Key: tuple.NewTupleKey("group:3", "member", "user:maria"),
+				},
+				{
+					Key: tuple.NewTupleKey("group:4", "member", "user:maria"),
+				},
+			},
+			readUsersetTuples: [][]*openfgav1.Tuple{
+				{
+					{
+						Key: tuple.NewTupleKey("group:1", "member", "group:3#member"),
+					},
+				},
+			},
+			expected: &ResolveCheckResponse{
+				Allowed: true,
+			},
+		},
+		{
+			name: "user_assigned_to_second_level_sub_group",
+			readStartingWithUserTuples: []*openfgav1.Tuple{
+				{
+					Key: tuple.NewTupleKey("group:3", "member", "user:maria"),
+				},
+				{
+					Key: tuple.NewTupleKey("group:4", "member", "user:maria"),
+				},
+			},
+			readUsersetTuples: [][]*openfgav1.Tuple{
+				{
+					{
+						Key: tuple.NewTupleKey("group:6", "member", "group:5#member"),
+					},
+				},
+				{
+					{
+						Key: tuple.NewTupleKey("group:5", "member", "group:3#member"),
+					},
+				},
+			},
+			expected: &ResolveCheckResponse{
+				Allowed: true,
+			},
+		},
+		{
+			name: "user_not_assigned_to_sub_group",
+			readStartingWithUserTuples: []*openfgav1.Tuple{
+				{
+					Key: tuple.NewTupleKey("group:3", "member", "user:maria"),
+				},
+				{
+					Key: tuple.NewTupleKey("group:4", "member", "user:maria"),
+				},
+			},
+			readUsersetTuples: [][]*openfgav1.Tuple{
+				{
+					{
+						Key: tuple.NewTupleKey("group:1", "member", "group:2#member"),
+					},
+				},
+				{},
+			},
+			expected: &ResolveCheckResponse{
+				Allowed: false,
+			},
+		},
+		{
+			name:                            "error_getting_tuple",
+			readStartingWithUserTuples:      []*openfgav1.Tuple{},
+			readStartingWithUserTuplesError: fmt.Errorf("mock error"),
+			readUsersetTuples: [][]*openfgav1.Tuple{
+				{
+					{
+						Key: tuple.NewTupleKey("group:1", "member", "group:2#member"),
+					},
+				},
+				{},
+			},
+			expected:      nil,
+			expectedError: fmt.Errorf("mock error"),
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+
+			ctrl := gomock.NewController(t)
+			defer ctrl.Finish()
+			storeID := ulid.Make().String()
+			ds := mocks.NewMockRelationshipTupleReader(ctrl)
+			ctx := context.Background()
+			ctx = storage.ContextWithRelationshipTupleReader(ctx, ds)
+
+			ds.EXPECT().ReadStartingWithUser(gomock.Any(), storeID, storage.ReadStartingWithUserFilter{
+				ObjectType: "group",
+				Relation:   "member",
+				UserFilter: []*openfgav1.ObjectRelation{{Object: "user:maria"}},
+				ObjectIDs:  nil,
+			}, gomock.Any()).MaxTimes(1).Return(storage.NewStaticTupleIterator(tt.readStartingWithUserTuples), tt.readStartingWithUserTuplesError)
+
+			for _, tuples := range tt.readUsersetTuples[1:] {
+				ds.EXPECT().ReadUsersetTuples(gomock.Any(), storeID, gomock.Any(), gomock.Any()).MaxTimes(1).Return(storage.NewStaticTupleIterator(tuples), tt.readUsersetTuplesError)
+			}
+			model := parser.MustTransformDSLToProto(`
+						model
+							schema 1.1
+
+						type user
+						type group
+							relations
+								define member: [user, group#member]
+`)
+
+			ts, err := typesystem.New(model)
+			require.NoError(t, err)
+			ctx = typesystem.ContextWithTypesystem(ctx, ts)
+
+			req := &ResolveCheckRequest{
+				StoreID:              storeID,
+				AuthorizationModelID: ulid.Make().String(),
+				TupleKey:             tuple.NewTupleKey("group:1", "member", "user:maria"),
+				RequestMetadata:      NewCheckRequestMetadata(),
+			}
+
+			checker := NewLocalChecker()
+			tupleKeys := make([]*openfgav1.TupleKey, 0, len(tt.readUsersetTuples[0]))
+			for _, t := range tt.readUsersetTuples[0] {
+				k := t.GetKey()
+				tupleKeys = append(tupleKeys, &openfgav1.TupleKey{
+					User:     k.GetUser(),
+					Relation: k.GetRelation(),
+					Object:   k.GetObject(),
+				})
+			}
+
+			result, err := checker.recursiveUsersetFastPath(ctx, req, storage.NewStaticTupleKeyIterator(tupleKeys))
+			require.Equal(t, tt.expectedError, err)
+			require.Equal(t, tt.expected.GetAllowed(), result.GetAllowed())
+			require.Equal(t, tt.expected.GetResolutionMetadata(), result.GetResolutionMetadata())
+		})
+	}
+
+	t.Run("resolution_depth_exceeded", func(t *testing.T) {
+		t.Parallel()
+		ctrl := gomock.NewController(t)
+		defer ctrl.Finish()
+
+		storeID := ulid.Make().String()
+		ds := mocks.NewMockRelationshipTupleReader(ctrl)
+
+		ctx := context.Background()
+		ctx = storage.ContextWithRelationshipTupleReader(ctx, ds)
+
+		ds.EXPECT().ReadStartingWithUser(gomock.Any(), storeID, gomock.Any(), gomock.Any()).MaxTimes(1).Return(
+			storage.NewStaticTupleIterator([]*openfgav1.Tuple{
+				{
+					Key: tuple.NewTupleKey("group:bad", "member", "user:maria"),
+				},
+			}), nil)
+
+		for i := 1; i < 26; i++ {
+			ds.EXPECT().ReadUsersetTuples(gomock.Any(), storeID, gomock.Any(), gomock.Any()).MaxTimes(1).Return(
+				storage.NewStaticTupleIterator([]*openfgav1.Tuple{
+					{
+						Key: tuple.NewTupleKey("group:"+strconv.Itoa(i+1), "member", "group:"+strconv.Itoa(i)+"#member"),
+					},
+				}), nil)
+		}
+		model := parser.MustTransformDSLToProto(`
+							model
+								schema 1.1
+
+							type user
+							type group
+								relations
+									define member: [user, group#member]
+
+			`)
+
+		ts, err := typesystem.New(model)
+		require.NoError(t, err)
+		ctx = typesystem.ContextWithTypesystem(ctx, ts)
+
+		req := &ResolveCheckRequest{
+			StoreID:              storeID,
+			AuthorizationModelID: ulid.Make().String(),
+			TupleKey:             tuple.NewTupleKey("group:1", "member", "user:maria"),
+			RequestMetadata:      NewCheckRequestMetadata(),
+		}
+
+		checker := NewLocalChecker()
+		tupleKeys := []*openfgav1.TupleKey{{Object: "group:1", Relation: "member", User: "group:0#member"}}
+
+		result, err := checker.recursiveUsersetFastPath(ctx, req, storage.NewStaticTupleKeyIterator(tupleKeys))
+		require.Nil(t, result)
+		require.Equal(t, ErrResolutionDepthExceeded, err)
+	})
+}
+
+func TestBuildRecursiveMapper(t *testing.T) {
+	mockController := gomock.NewController(t)
+	defer mockController.Finish()
+
+	mockDatastore := mocks.NewMockOpenFGADatastore(mockController)
+
+	storeID := ulid.Make().String()
+	ctx := context.Background()
+	ctx = storage.ContextWithRelationshipTupleReader(ctx, mockDatastore)
+
+	model := testutils.MustTransformDSLToProtoWithID(`
+			model
+				schema 1.1
+			type user
+			type group
+				relations
+					define member: [user]`)
+	ts, err := typesystem.New(model)
+	require.NoError(t, err)
+
+	ctx = typesystem.ContextWithTypesystem(ctx, ts)
+	checker := NewLocalChecker()
+
+	t.Run("nested_userset", func(t *testing.T) {
+		mockDatastore.EXPECT().ReadUsersetTuples(ctx, storeID, storage.ReadUsersetTuplesFilter{
+			Object:   "document:1",
+			Relation: "viewer",
+			AllowedUserTypeRestrictions: []*openfgav1.RelationReference{
+				typesystem.DirectRelationReference("group", "member"),
+			},
+		}, storage.ReadUsersetTuplesOptions{
+			Consistency: storage.ConsistencyOptions{
+				Preference: openfgav1.ConsistencyPreference_HIGHER_CONSISTENCY,
+			},
+		}).Times(1)
+
+		mapping := &recursiveMapping{
+			kind: UsersetKind,
+			allowedUserTypeRestrictions: []*openfgav1.RelationReference{
+				typesystem.DirectRelationReference("group", "member"),
+			},
+		}
+		res, err := checker.buildRecursiveMapper(ctx, &ResolveCheckRequest{
+			StoreID:     storeID,
+			TupleKey:    tuple.NewTupleKey("document:1", "viewer", "user:maria"),
+			Context:     testutils.MustNewStruct(t, map[string]interface{}{"x": "2"}),
+			Consistency: openfgav1.ConsistencyPreference_HIGHER_CONSISTENCY,
+		}, mapping)
+		require.NoError(t, err)
+		_, ok := res.(*UsersetMapper)
+		require.True(t, ok)
+	})
+
+	t.Run("nested_ttu", func(t *testing.T) {
+		mockDatastore.EXPECT().Read(ctx, storeID, tuple.NewTupleKey("document:1", "parent", ""), storage.ReadOptions{
+			Consistency: storage.ConsistencyOptions{
+				Preference: openfgav1.ConsistencyPreference_HIGHER_CONSISTENCY,
+			},
+		}).Times(1)
+
+		mapping := &recursiveMapping{
+			tuplesetRelation: "parent",
+			kind:             TTUKind,
+		}
+		res, err := checker.buildRecursiveMapper(ctx, &ResolveCheckRequest{
+			StoreID:     storeID,
+			TupleKey:    tuple.NewTupleKey("document:1", "viewer", "user:maria"),
+			Context:     testutils.MustNewStruct(t, map[string]interface{}{"x": "2"}),
+			Consistency: openfgav1.ConsistencyPreference_HIGHER_CONSISTENCY,
+		}, mapping)
+		require.NoError(t, err)
+		_, ok := res.(*TTUMapper)
+		require.True(t, ok)
 	})
 }
